Remove AssetBundleGroupSharedAsset
