--- Packages/UGF.Module.AssetBundles/Editor/AssetBundleGroupSharedAssetEditor.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: bac9eeb0f7444ee9ae62e6963c0d2de7
-timeCreated: 1667064094
