--- Packages/UGF.Module.AssetBundles/Runtime/AssetBundleGroupSharedAsset.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 83ab82ecb70040e2a7acd28f93ae0a94
-timeCreated: 1667063989
